requirements.txt
streamlit
re
python-docx
nlkt
ngrams
fpdf